+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ### Added
-- `SpriteIcon`
+- `SpriteIcon` (use type `"sprite"`)
+- Added an `ExtendedLanguageProvider` that allows you to read existing json files when generating translations
 
 ### Changed
 - Renamed some stuff